tests/requirements.txt
flake8==4.0.1
mypy==0.971
pytest==7.1.2
pytest-mock==3.8.2
pytest-cov==3.0.0
pytest-xprocess==0.19.0
seleniumbase==3.5.8
